WEB-70: [refactor] split 'alma.py' into parts
- abstract 'Service'
- record_utils for marcxml/ltree
- 'bibs' service containing all of the previous ones
