+++ src/services/__init__.py
@@ -1,0 +1,5 @@
+
+# helpful reused variables
+HTTP_TOO_MANY_REQUESTS = 429
+HTTP_BAD_REQUEST = 400
+CONTENT_TYPE_XML = {'Content-Type': 'application/xml'}
